perf: re video

+++ core/pipeline.py
@@ -1,17 +1,17 @@
-from core.imageupscaler import ImageUpscaler
 import os
 import uuid
-from tools.utils import get_model_list, timer
-# from threading import Thread
-from tools.tpu_utils import clean_tpu_memory
 import cv2
 import numpy as np
-from plugin.bgremove import Bgremover, Bgremover2
 
+from tools.utils import get_model_list, timer, fuse_audio_with_ffmpeg
+from tools.tpu_utils import clean_tpu_memory
+from core.imageupscaler import ImageUpscaler, ImageUpscaler2
+from plugin.bgremove import Bgremover, Bgremover2
 
 image_model_list = get_model_list()
 video_model_list = get_model_list("video")
 image_upscaler = ImageUpscaler()
+image_upscaler2 = ImageUpscaler2()
 bger = Bgremover()
 bger2 = Bgremover2()
 
@@ -50,11 +50,11 @@ def image_pipeline(input, model, face_enhance=None, background_remove=None, outp
         img = input
     if background_remove == 0:
         image_upscaler.change_model(model, face_enhance_name="None", bg_helper=True)
-        bg_hlepr_img = image_upscaler(img, face_enhance="None", bg_upscale=True)
+        bg_hlepr_img = image_upscaler.forward(img, face_enhance="None", bg_upscale=True)
         bger.init_model()
         _, mask = bger.forward(bg_hlepr_img)
         image_upscaler.change_model(model, face_enhance_name=face_enhance, bg_helper=False)
-        res = image_upscaler(img, face_enhance)
+        res = image_upscaler.forward(img, face_enhance)
         new_mask = cv2.resize(mask, (res.shape[1], res.shape[0]), interpolation=cv2.INTER_LINEAR)
         res = cv2.cvtColor(res, cv2.COLOR_RGB2BGRA)
         res = res[:,:,[2,1,0,3]]
@@ -68,7 +68,7 @@ def image_pipeline(input, model, face_enhance=None, background_remove=None, outp
     elif background_remove == 2:
         # only upscale
         image_upscaler.change_model(model, face_enhance, bg_helper=False)
-        res = image_upscaler(img, face_enhance, bg_upscale=False)
+        res = image_upscaler.forward(img, face_enhance, bg_upscale=False)
 
     if save:
         if output_path is None:
@@ -87,8 +87,6 @@ def image_pipeline(input, model, face_enhance=None, background_remove=None, outp
 @timer
 def video_pipeline(input, model, face_enhance=None, background_remove=None, thread=1, audio=None, output_path=None, save=True):
     clean_tpu_memory([image_upscaler, bger])
-    print(type(input))
-    print(input)
     if not isinstance(input, str) or input is None:
         print("Please upload your video")
         return "Please upload your video", None
@@ -106,8 +104,10 @@ def video_pipeline(input, model, face_enhance=None, background_remove=None, thre
     cap = cv2.VideoCapture(input)
     fps = cap.get(cv2.CAP_PROP_FPS)
     cap.release()
+    print("**************")
     os.system(f'ffmpeg_ubuntu -i {input} -qscale:v 1 -qmin 1 -qmax 1 -vsync 0  ./temp_frames/frame%d.png')
     print("**************")
+
     if background_remove == 0:
         pass
 
@@ -118,7 +118,10 @@ def video_pipeline(input, model, face_enhance=None, background_remove=None, thre
         pass
 
     elif background_remove == 2:
-        pass
+        # only upscale
+        image_upscaler2.change_model(model, face_enhance, thread_num=thread)
+        image_upscaler2.forward('./temp_frames')
+
 
     if save:
         if output_path is None:
@@ -127,14 +130,17 @@ def video_pipeline(input, model, face_enhance=None, background_remove=None, thre
             output_path = './result/video/out_{}.mp4'.format(uuid.uuid4())
         print("merge")
         os.system(f'ffmpeg_ubuntu -framerate {fps} -i ./temp_res_frames/frame%d.png -c:v libx264 -pix_fmt yuv420p -y {output_path}')
+        if audio:
+            output_path = fuse_audio_with_ffmpeg(input,output_path)
     clean_cache_file()
     # os.remove(input)
     return "Video Process Success, you can find the output in {} or click the download icon".format(output_path), output_path
 
 
-def clean_cache_file():
+def clean_cache_file(mask=False):
     temp_frames = os.listdir("./temp_frames")
     for i in temp_frames:
         os.remove(os.path.join("./temp_frames/{}".format(i)))
-        os.remove(os.path.join("./temp_mask/{}".format(i)))
+        if mask:
+            os.remove(os.path.join("./temp_mask/{}".format(i)))
         os.remove(os.path.join("./temp_res_frames/{}".format(i)))
